Removed viewer from upstream tesseract lib.
I also added the "-all_load" flag to the Template Framework
Project target as well as the TestsProject target and I can
confirm that both build successfully and all the tests run okay.

But I didn't add it to the TestsProjectTests target because I
wasn't sure it belongs there. Feel free to add it there as well
if you think it's needed.